project/backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.4
pytz==2023.3
